requirements.txt
telethon==1.34.0
boto3==1.35.36
botocore==1.35.36
python-dotenv==1.0.1
tenacity==9.0.0
requests==2.32.3
